Move default constants to respective _types.go files and delete defaults.go
